Change id columns from int to long

+++ sql/001_create_tables.sql
@@ -1,16 +1,16 @@
 USE royals_of_britain;
 
 CREATE TABLE `houses` (
-    `id` int PRIMARY KEY AUTO_INCREMENT,
+    `id` bigint(20) PRIMARY KEY AUTO_INCREMENT,
     `name` varchar(255) NOT NULL
 );
 CREATE TABLE `kings` (
-    `id` int AUTO_INCREMENT PRIMARY KEY,
+    `id` bigint(20) AUTO_INCREMENT PRIMARY KEY,
     `name` varchar(255) NOT NULL,
     `year_of_birth` int,
     `year_of_death` int,
-    `father_id` int,
-    `house_id` int,
+    `father_id` bigint(20),
+    `house_id` bigint(20),
     FOREIGN KEY (house_id) REFERENCES houses(id)
 );
